Build and test matrix without warnings
