+++ gneiss-mqtt/src/client.rs
@@ -1003,102 +1003,6 @@ impl MqttClientImpl {
 /// Module containing shared types used to wait on the emission of specific client events from a client.
 ///
 /// Unstable; may be refactored or made internal.
-pub mod waiter {
-
-    use super::*;
-
-    /// Simple C-style enum whose entries match ClientEvent.  Useful for coarse matching against event types when we don't
-    /// need to dig into the variant's internal data.
-    #[derive(Debug, Eq, PartialEq, Copy, Clone)]
-    #[non_exhaustive]
-    pub enum ClientEventType {
-
-        /// Corresponds to a ClientEvent::ConnectionAttempt, the client event emitted every time the client tries
-        /// to establish a connection
-        ConnectionAttempt,
-
-        /// Corresponds to a ClientEvent::ConnectionSuccess, the client event emitted every time the client successfully
-        /// connects to an MQTT broker (received a successful CONNACK)
-        ConnectionSuccess,
-
-        /// Corresponds to a ClientEvent::ConnectionFailure, the client event emitted every time a client's connection
-        /// attempt results in some kind of failure
-        ConnectionFailure,
-
-        /// Corresponds to a ClientEvent::Disconnection, the client event emitted every time a successfully connected
-        /// client has its connection closed for any reason
-        Disconnection,
-
-        /// Corresponds to a ClientEvent::Stopped, the client event emitted every time a previously running client
-        /// settles into the Stopped state by user request.
-        Stopped,
-
-        /// Corresponds to a ClientEvent::PublishReceived, the client event emitted every time a Publish packet
-        /// is received
-        PublishReceived,
-    }
-
-    /// Checks if a ClientEvent matches a ClientEventType
-    pub fn client_event_matches(event: &Arc<ClientEvent>, event_type: ClientEventType) -> bool {
-        match **event {
-            ClientEvent::ConnectionAttempt(_) => { event_type == ClientEventType::ConnectionAttempt }
-            ClientEvent::ConnectionSuccess(_) => { event_type == ClientEventType::ConnectionSuccess }
-            ClientEvent::ConnectionFailure(_) => { event_type == ClientEventType::ConnectionFailure }
-            ClientEvent::Disconnection(_) => { event_type == ClientEventType::Disconnection }
-            ClientEvent::Stopped(_) => { event_type == ClientEventType::Stopped }
-            ClientEvent::PublishReceived(_) => { event_type == ClientEventType::PublishReceived }
-        }
-    }
-
-    /// Filter function type used to create complex waiters.  Only events that pass the filter check will be
-    /// passed on to the user at the conclusion of the wait() call.
-    pub type ClientEventPredicate = dyn Fn(&Arc<ClientEvent>) -> bool + Send + Sync;
-
-    /// Enum controlling how the waiter should filter client events.
-    pub enum ClientEventWaitType {
-
-        /// Filter all client events that do not match this variant's type
-        Type(ClientEventType),
-
-        /// Filter all client events that the predicate filter function returns false for
-        Predicate(Box<ClientEventPredicate>)
-    }
-
-    /// Configuration options for waiter creation
-    pub struct ClientEventWaiterOptions {
-
-        /// How the waiter should filter client events
-        pub wait_type: ClientEventWaitType,
-    }
-
-    #[derive(Clone)]
-    /// Timestamped client event record
-    pub struct ClientEventRecord {
-
-        /// The event emitted by the client
-        pub event : Arc<ClientEvent>,
-
-        /// What time the event occurred at
-        pub timestamp: Instant
-    }
-
-    /// Result type for calling wait() on an async client event waiter
-    pub type ClientEventWaitFuture = dyn Future<Output = MqttResult<Vec<ClientEventRecord>>> + Send;
-
-    /// Trait API for waiting for a set of client events to be emitted by an async client
-    pub trait AsyncClientEventWaiter {
-
-        /// Waits for and returns an event sequence that matches the configuration the waiter was created with
-        fn wait(self) -> Pin<Box<ClientEventWaitFuture>>;
-    }
-
-    /// Trait API for waiting for a set of client events to be emitted by a synchronous client
-    pub trait SyncClientEventWaiter {
-
-        /// Waits for and returns an event sequence that matches the configuration the waiter was created with
-        fn wait(self) -> MqttResult<Vec<ClientEventRecord>>;
-    }
-}
 
 /// An async network client that functions as a thin wrapper over the MQTT5 protocol.
 ///
@@ -1194,13 +1098,11 @@ pub struct SyncResultReceiver<T> {
     result_signal: Arc<Condvar>
 }
 
-#[allow(dead_code)]
 pub(crate) struct SyncResultSender<T> {
     result_lock: Arc<Mutex<Option<T>>>,
     result_signal: Arc<Condvar>
 }
 
-#[allow(dead_code)]
 impl<T> Clone for SyncResultSender<T> {
     fn clone(&self) -> Self {
         SyncResultSender {
@@ -1210,7 +1112,6 @@ impl<T> Clone for SyncResultSender<T> {
     }
 }
 
-#[allow(dead_code)]
 impl<T> SyncResultSender<T> {
 
     pub(crate) fn new(result_lock: Arc<Mutex<Option<T>>>, result_signal: Arc<Condvar>) -> SyncResultSender<T> {
@@ -1220,6 +1121,7 @@ impl<T> SyncResultSender<T> {
         }
     }
 
+    #[cfg_attr(not(feature="threaded"), allow(dead_code))]
     pub(crate) fn apply(&self, value: T) {
         let mut current_value = self.result_lock.lock().unwrap();
 
@@ -1235,7 +1137,6 @@ impl<T> SyncResultSender<T> {
 
 impl<T> SyncResultReceiver<T> {
 
-    #[allow(dead_code)]
     pub(crate) fn new(result_lock: Arc<Mutex<Option<T>>>, result_signal: Arc<Condvar>) -> SyncResultReceiver<T> {
         SyncResultReceiver {
             result_lock,
@@ -1265,7 +1166,7 @@ impl<T> SyncResultReceiver<T> {
     }
 }
 
-#[allow(dead_code)]
+#[cfg_attr(not(feature="threaded"), allow(dead_code))]
 pub(crate) fn new_sync_result_pair<T>() -> (SyncResultReceiver<T>, SyncResultSender<T>) {
     let lock = Arc::new(Mutex::new(None));
     let signal = Arc::new(Condvar::new());
