+++ gneiss-mqtt/src/lib.rs
@@ -207,6 +207,9 @@ in the README for more details.
 
 #![warn(missing_docs)]
 
+#![cfg_attr(not(any(feature = "tokio", feature = "threaded")), allow(dead_code))]
+#![cfg_attr(all(feature = "testing", not(test)), allow(dead_code, unused_imports, unused_macros))]
+
 pub mod alias;
 pub mod client;
 pub mod config;
