+++ gneiss-mqtt/src/testing/mod.rs
@@ -3,7 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
+#[cfg(any(feature = "tokio", feature = "threaded"))]
 pub(crate) mod integration;
 pub(crate) mod protocol;
 pub(crate) mod mock_server;
+pub mod waiter;
